Render children route on Producers base component
