+++ src/containers/Group/Producers/index.js
@@ -1,2 +1,3 @@
 export GroupProducersBase from './Base';
+export GroupProducersList from './List';
 export GroupProducersDetails from './Details';
