+++ src/containers/index.js
@@ -5,7 +5,7 @@ export { Create as Signup, Complete as SignupComplete } from './Signup';
 export InvitationComplete from './Invitation/Complete';
 export GroupsList from './Groups';
 export { GroupBase, GroupMembers } from './Group';
-export { GroupProducersBase, GroupProducersDetails } from './Group/Producers';
+export { GroupProducersBase, GroupProducersList, GroupProducersDetails } from './Group/Producers';
 export { OnboardingCreateGroup, OnboardingInvitations } from './Onboarding';
 export NotFound from './NotFound/NotFound';
