+++ src/routes.js
@@ -6,21 +6,22 @@ import { isLoaded as isAuthLoaded, load as loadAuth } from 'redux/modules/auth';
 import { load as loadMemberships } from 'redux/modules/groups/memberships';
 
 import {
-    App,
-    Home,
-    Login,
-    Signup,
-    SignupComplete,
-    GroupsList,
-    GroupBase,
-    GroupMembers,
-    GroupProducersBase,
-    GroupProducersDetails,
-    OnboardingCreateGroup,
-    OnboardingInvitations,
-    InvitationComplete,
-    NotFound,
-  } from 'containers';
+  App,
+  Home,
+  Login,
+  Signup,
+  SignupComplete,
+  GroupsList,
+  GroupBase,
+  GroupMembers,
+  GroupProducersBase,
+  GroupProducersList,
+  GroupProducersDetails,
+  OnboardingCreateGroup,
+  OnboardingInvitations,
+  InvitationComplete,
+  NotFound,
+} from 'containers';
 
 /**
  * Go to group detail if user only has 1 membership
@@ -126,6 +127,7 @@ export default (store) => {
             <IndexRoute component={GroupMembers}/>
             <Route path="members" component={GroupMembers}/>
             <Route path="producers" component={GroupProducersBase}>
+              <IndexRoute component={GroupProducersList} />
               <Route path=":producer_id" component={GroupProducersDetails}/>
             </Route>
           </Route>
